# sql: don't copy ColumnFamilyDescriptor and IndexDescriptor by value in loops