--- a/pkg/sql/row/writer.go
+++ b/pkg/sql/row/writer.go
@@ -87,7 +87,8 @@ func prepareInsertOrUpdateBatch(
 	putFn func(ctx context.Context, b putter, key *roachpb.Key, value *roachpb.Value, traceKV bool),
 	overwrite, traceKV bool,
 ) ([]byte, error) {
-	for i, family := range helper.TableDesc.Families {
+	for i := range helper.TableDesc.Families {
+		family := &helper.TableDesc.Families[i]
 		update := false
 		for _, colID := range family.ColumnIDs {
 			if _, ok := marshaledColIDMapping[colID]; ok {