`ColumnFamilyDescriptor`s are 80 bytes and `IndexDescriptor`s are 392 bytes, so they're too big to be copying by value when it's trivial to avoid doing so.

I was originally prompted to look into this when I saw the copy in `prepareInsertOrUpdateBatch` account for .59% of a CPU profile.

#### Performance diff (excluding the first commit):
```
name                       old time/op    new time/op    delta
KV/Insert/SQL/rows=1000-4    24.0ms ± 3%    23.6ms ± 2%  -1.88%  (p=0.027 n=10+8)

name                       old alloc/op   new alloc/op   delta
KV/Insert/SQL/rows=1000-4    4.89MB ± 0%    4.89MB ± 0%    ~     (p=0.796 n=10+10)

name                       old allocs/op  new allocs/op  delta
KV/Insert/SQL/rows=1000-4     26.8k ± 0%     26.8k ± 0%    ~     (p=0.838 n=10+10)
```