@nvanbenschoten 2019-05-23
`raceTransport` caught what it considers a "data race" because parallel commits had a code path that would mutate a request after it had returned a response. Fixed in a new commit.

bors r+